WIP. Add BirliContext and control progress bars.
This commit adds a new struct, BirliContext, to do Birli-related tasks.
Personally, this is tidier than having "dangling" functions around the
code. But more importantly, if/when Birli supports things like "Levine
gains", it's going to be *much* tidier to instantiate a struct with your
choice of gains once, then every call to "context_to_jones_array". In
addition, the way I've designed BirliContext allows it to transparently
use CorrelatorContext methods and members, so one could think of it as
an augmented mwalib context.

I'm not super happy with the name BirliContext - suggestions?

More/all functions have to be moved into the impl BirliContext block.

In theory, this commit also allows progress bars to be turned off, but
right now they still render as ugly grey blocks in my terminal. Not sure
what's going on.

I'll be happy to hack away at this as time allows, but of course I'd be
happy to get your feedback.