Overhaul mwaf files.
Fixes https://github.com/MWATelescope/Birli/issues/16

MWA flag files ("mwaf" files) were historically bespoke but potentially
extremely useful for diagnostics. This commit modernises the file format
by adding and changing a few header keys:

Add:
OBSID: The MWA observation ID,
GPSSTART: The centroid time associated with the first timestep in the
flags,
SOFTWARE: The name and version of the software used to write the mwaf
file,
AO_VER: AOFlagger version,
AO_STRAT: AOFlagger strategy file used (as a fits long string),

as well as a new fits table in the "OCCUPANCY" HDU: the first
column ("Index") is the channel index for this mwaf file (starting from
0), the second column ("Count") is the flag count associated with this
channel, and the third column ("Occupancy") is the flag occupancy for
this channel (i.e. the number of flags for this channel divided by the
total number of rows, as a value between 0 and 1).

Remove header keys: GPSTIME (now OBSID), COTVER (now SOFTWARE),
COTVDATE.

In addition to the above, mwaf files are now written out in parallel,
and any existing mwaf files are overwritten when running Birli.

Remove crossbeam-channel as a dependency, specify the crossbeam-utils
dep version more tightly, and move lazy_static to the dev-dependencies.