WIP. Use even chunk sizes.
WIP: Needs to use an integer number of averaged timesteps

This has the advantage of using less memory and having more consistent
flagging.

e.g. If we have a 400 GB observation, and we're requested to use no more
than 300 GB (i.e. --max-memory=300), then the previous behaviour was to
use 300 GB in the first chunk of work, then 100 GB in the next. This
commit opts to use 200 GB in each chunk instead (i.e. make the chunk
sizes more even), because (1) this uses less memory and is therefore
less likely to fail, and (2) should have more consistent flagging
behaviour.